+++ app/src/main/java/com/codeclinic/agent/model/user/UserRoleModel.java
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName;
 public class UserRoleModel {
     @SerializedName("dateTimeCreated")
     @Expose
-    private Integer dateTimeCreated;
+    private long dateTimeCreated;
     @SerializedName("name")
     @Expose
     private String name;
@@ -17,11 +17,11 @@ public class UserRoleModel {
     @Expose
     private Integer id;
 
-    public Integer getDateTimeCreated() {
+    public long getDateTimeCreated() {
         return dateTimeCreated;
     }
 
-    public void setDateTimeCreated(Integer dateTimeCreated) {
+    public void setDateTimeCreated(long dateTimeCreated) {
         this.dateTimeCreated = dateTimeCreated;
     }
