+++ app/src/main/java/com/codeclinic/agent/model/user/UserTenetModel.java
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName;
 public class UserTenetModel {
     @SerializedName("dateTimeCreated")
     @Expose
-    private Integer dateTimeCreated;
+    private long dateTimeCreated;
     @SerializedName("id")
     @Expose
     private Integer id;
@@ -29,11 +29,11 @@ public class UserTenetModel {
     @Expose
     private UserRealm realm;
 
-    public Integer getDateTimeCreated() {
+    public long getDateTimeCreated() {
         return dateTimeCreated;
     }
 
-    public void setDateTimeCreated(Integer dateTimeCreated) {
+    public void setDateTimeCreated(long dateTimeCreated) {
         this.dateTimeCreated = dateTimeCreated;
     }
